+++ theme/static/scripts/model_builder_main.js
@@ -18,7 +18,6 @@ function initSortableObjectCards() {
 function initModelBuilderMain() {
     initLeaderLines();
     initSortableObjectCards();
-    initSidePanel();
     initHammer();
     resizeSystemNameHeader();
 }